rename hook files to conform to file naming convention
